1.0.2
Final changes and bug fixes. Actually had things to do but the time has ended so if nothing else happens this is the final version of the game. Had a major fix in the gameplayMethods.getName, that was not good, now it should be. Some smaller repairs, leftover files, things been deleted.